Make Upload layers their own components
