+++ src/components/MediaCrud/index.js
@@ -1,11 +1,9 @@
 import MediaCrudContainer from 'components/MediaCrud/MediaCrudContainer'
 import MediaCrudData from 'components/MediaCrud/MediaCrudData'
 import SelectMedia from 'components/MediaCrud/SelectMedia'
-import UploadMedia from 'components/MediaCrud/UploadMedia'
 
 export default {
   Container: MediaCrudContainer,
   Data: MediaCrudData,
   SelectMedia,
-  UploadMedia,
 }
