+++ src/components/UploadAudio/index.js
@@ -1,0 +1,3 @@
+import UploadAudio from 'components/UploadAudio/UploadAudio'
+
+export default UploadAudio
