+++ src/components/UploadVisualMedia/index.js
@@ -1,0 +1,3 @@
+import UploadVisualMedia from 'components/UploadVisualMedia/UploadVisualMedia'
+
+export default UploadVisualMedia
